+++ src/appengine_magic/core.clj
@@ -1,4 +1,5 @@
-(ns appengine-magic.core)
+(ns appengine-magic.core
+  (:import com.google.apphosting.api.ApiProxy))
 
 
 (defn in-appengine-interactive-mode? []
@@ -12,12 +13,30 @@
       false)))
 
 
+(defn open-resource-stream [resource-name]
+  (-> (clojure.lang.RT/baseLoader) (.getResourceAsStream resource-name)))
+
+
 (defn appengine-environment-type []
   (let [env-property (System/getProperty "com.google.appengine.runtime.environment")]
     (cond
      (nil? env-property) :interactive
      (= env-property "Development") :dev-appserver
      (= env-property "Production") :production)))
+
+
+(defn appengine-app-id []
+  (try
+    (-> (ApiProxy/getCurrentEnvironment) .getAppId)
+    (catch NullPointerException npe
+      (throw (RuntimeException. "the server must be running" npe)))))
+
+
+(defn appengine-app-version []
+  (try
+    (-> (ApiProxy/getCurrentEnvironment) .getVersionId)
+    (catch NullPointerException npe
+      (throw (RuntimeException. "the server must be running" npe)))))
 
 
 (if (in-appengine-interactive-mode?)