+++ src/appengine_magic/core_google.clj
@@ -3,9 +3,12 @@
 (import '[java.io File FileInputStream BufferedInputStream])
 
 
-(defn open-resource-stream [resource-name]
-  (let [f (File. resource-name)]
-    (BufferedInputStream. (FileInputStream. resource-name))))
+(defn appengine-base-url [& {:keys [https?] :or {https? false}}]
+  (when (= :dev-appserver (appengine-environment-type))
+    (throw (RuntimeException.
+            "appengine-magic.core/appengine-base-url not supported in dev-appserver.sh")))
+  (str (if https? "https" "http")
+       "://" (appengine-app-id) ".appspot.com"))
 
 
 (defmacro def-appengine-app [app-var-name handler & [args]]