+++ src/appengine_magic/servlet.clj
@@ -1,20 +1,13 @@
 ;;; This code is adapted from Ring (http://github.com/mmcgrana/ring).
 ;;;
-;;; Required changes from Ring:
-;;;   1. Remove dependencies which use Java classes blacklisted in App Engine.
-;;;   2. Include raw cookie data retrieved using the servlet API. This does not
-;;;      preclude using Ring's cookie middleware, but is required separately:
-;;;      App Engine services do not always use standard-compliant cookies which
-;;;      the Ring middleware parses correctly.
+;;; Required change from Ring: removed dependencies which use Java classes
+;;; blacklisted in App Engine.
 
 
 (ns appengine-magic.servlet
   (:use [appengine-magic.utils :only [copy-stream]])
   (:import [java.io File FileInputStream InputStream ByteArrayInputStream OutputStream]
            [javax.servlet.http HttpServlet HttpServletRequest HttpServletResponse]))
-
-
-(defrecord Cookie [comment domain max-age name path secure value version])
 
 
 (defn- get-headers [^HttpServletRequest request]
@@ -24,21 +17,6 @@
           (enumeration-seq (.getHeaderNames request))))
 
 
-(defn- get-cookies [^HttpServletRequest request]
-  (reduce (fn [cookies ^javax.servlet.http.Cookie new-cookie]
-            (assoc cookies (.getName new-cookie)
-                   (Cookie. (.getComment new-cookie)
-                            (.getDomain new-cookie)
-                            (.getMaxAge new-cookie)
-                            (.getName new-cookie)
-                            (.getPath new-cookie)
-                            (.getSecure new-cookie)
-                            (.getValue new-cookie)
-                            (.getVersion new-cookie))))
-          {}
-          (.getCookies request)))
-
-
 (defn- make-request-map [^HttpServlet servlet
                          ^HttpServletRequest request
                          ^HttpServletResponse response]
@@ -46,7 +24,6 @@
    :response           response
    :request            request
    :servlet-context    (.getServletContext servlet)
-   :servlet-cookies    (get-cookies request)
    :server-port        (.getServerPort request)
    :server-name        (.getServerName request)
    :remote-addr        (.getRemoteAddr request)