+++ src/leiningen/appengine_prepare.clj
@@ -1,7 +1,8 @@
 (ns leiningen.appengine-prepare
   "Prepares a the Google App Engine application for deployment."
   (:use appengine-magic.utils)
-  (:require leiningen.compile leiningen.jar leiningen.util.file lancet)
+  (:require leiningen.compile leiningen.jar leiningen.util.file
+            [lancet.core :as lancet])
   (:import java.io.File))
 
 
@@ -9,35 +10,44 @@
   (let [prj-application (or (:appengine-application project) (:name project))
         prj-display-name (or (:appengine-display-name project) (:name project))
         prj-servlet (or (:appengine-entry-servlet project) "app_servlet")
-        resources-dir (File. (:resources-path project))
+        war-dir (File. (or (:appengine-app-war-root project) "war"))
         lib-dir (File. (:library-path project))
         lib-dev-dir (File. lib-dir "dev")
-        WEB-INF-dir (File. resources-dir "WEB-INF")
-        target-lib-dir (File. WEB-INF-dir "lib")
-        target-app-jar (File. target-lib-dir (str prj-application ".jar"))
+        web-inf-dir (File. war-dir "WEB-INF")
+        target-lib-dir (File. web-inf-dir "lib")
         compile-path (File. (:compile-path project))
         compile-path-exists? (.isDirectory compile-path)
         compile-path-empty? (= 0 (-> compile-path .list seq count))]
     (println "preparing App Engine application" prj-display-name "for deployment")
-    ;; compile all
-    (let [project-with-aot (if (contains? project :aot)
-                               project
-                               ;; For projects which do not define any
-                               ;; namespaces for AOT compilation, just include
-                               ;; the entry-point servlet.
-                               (assoc project
-                                 :aot [(symbol (format "%s.%s"
-                                                       (_dash prj-application) prj-servlet))]))]
-      (leiningen.compile/compile project-with-aot))
-    ;; delete existing content of target lib/
-    (lancet/delete {:dir (.getPath target-lib-dir)})
-    ;; prepare destination lib/ directory
-    (lancet/mkdir {:dir target-lib-dir})
-    ;; make a jar of the compiled app, and put it in WEB-INF/lib
-    (leiningen.jar/jar project)
-    (lancet/move {:file (leiningen.jar/get-jar-filename project)
-                  :todir (.getPath target-lib-dir)})
-    ;; projects which do not normally use AOT may need some cleanup
+    ;; compile all; when successful (status is 0), continue to prepare
+    (when (= 0 (leiningen.compile/compile (if (contains? project :aot)
+                                              project
+                                              (assoc project
+                                                :keep-non-project-classes true
+                                                :aot [(symbol (format "%s.%s"
+                                                                      (_dash prj-application)
+                                                                      prj-servlet))]))))
+      ;; delete existing content of target lib/
+      (lancet/delete {:dir (.getPath target-lib-dir)})
+      ;; prepare destination lib/ directory
+      (lancet/mkdir {:dir target-lib-dir})
+      ;; make a jar of the compiled app, and put it in WEB-INF/lib
+      (leiningen.jar/jar (merge project
+                                {:omit-source true
+                                 :jar-exclusions [#"^WEB-INF/appengine-generated.*$"]}))
+      (lancet/move {:file (leiningen.jar/get-jar-filename project)
+                    :todir (.getPath target-lib-dir)})
+      ;; copy important dependencies into WEB-INF/lib
+      (lancet/copy {:todir (.getPath target-lib-dir)}
+                   (lancet/fileset {:dir lib-dir :includes "*" :excludes "dev"}))
+      (lancet/copy {:todir (.getPath target-lib-dir)}
+                   (lancet/fileset
+                    {:dir lib-dev-dir
+                     :includes (str "appengine-magic*,ring-core*,servlet-api*,"
+                                    "commons-io*,commons-codec*,commons-fileupload*,"
+                                    "appengine-api-1.0-sdk*,appengine-api-labs*")})))
+    ;; Projects which do not normally use AOT may need some cleanup. This should
+    ;; happen regardless of compilation success or failure.
     (when-not (contains? project :aot)
       (cond
        ;; never had a classes/ directory; unlikely with Leiningen
@@ -47,11 +57,4 @@
        compile-path-empty?
        (doseq [entry-name (.list compile-path)]
          (let [entry (File. compile-path entry-name)]
-           (leiningen.util.file/delete-file-recursively entry true)))))
-    ;; copy important dependencies into WEB-INF/lib
-    (lancet/copy {:todir (.getPath target-lib-dir)}
-                 (lancet/fileset {:dir lib-dir :includes "*" :excludes "dev"}))
-    (lancet/copy {:todir (.getPath target-lib-dir)}
-                 (lancet/fileset
-                  {:dir lib-dev-dir
-                   :includes "appengine-magic*,ring-core*,appengine-api-1.0-sdk*,appengine-api-labs*"}))))
+           (leiningen.util.file/delete-file-recursively entry true)))))))